requirements.txt
# Amazon Smart Agent - Requirements
# Core dependencies
python-dotenv>=0.19.0
requests>=2.26.0
beautifulsoup4>=4.10.0
lxml>=4.6.3
selenium>=4.0.0
webdriver-manager>=3.5.2

# Amazon Product API
python-amazon-paapi>=4.1.0
bottlenose>=1.1.8

# Database
sqlalchemy>=1.4.23
alembic>=1.7.3

# Telegram Bot
python-telegram-bot>=13.7

# Web Dashboard
flask>=2.0.1
flask-restful>=0.3.9
flask-cors>=3.0.10

# CLI Tool
click>=8.0.1
rich>=10.9.0
tabulate>=0.8.9

# Utilities
pandas>=1.3.2
numpy>=1.21.2
tqdm>=4.62.2
loguru>=0.5.3

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1
